+++ sdk/storage/Azure.Storage.DataMovement.Blobs/api/Azure.Storage.DataMovement.Blobs.net6.0.cs
@@ -13,10 +13,8 @@ namespace Azure.Storage.DataMovement.Blobs
     public partial class AppendBlobStorageResource : Azure.Storage.DataMovement.StorageResourceItem
     {
         public AppendBlobStorageResource(Azure.Storage.Blobs.Specialized.AppendBlobClient blobClient, Azure.Storage.DataMovement.Blobs.AppendBlobStorageResourceOptions options = null) { }
-        protected override bool CanProduceUri { get { throw null; } }
         protected override long? Length { get { throw null; } }
         protected override long MaxChunkSize { get { throw null; } }
-        public override string Path { get { throw null; } }
         protected override string ResourceId { get { throw null; } }
         protected override Azure.Storage.DataMovement.DataTransferOrder TransferType { get { throw null; } }
         public override System.Uri Uri { get { throw null; } }
@@ -66,8 +64,6 @@ namespace Azure.Storage.DataMovement.Blobs
     public partial class BlobStorageResourceContainer : Azure.Storage.DataMovement.StorageResourceContainer
     {
         public BlobStorageResourceContainer(Azure.Storage.Blobs.BlobContainerClient blobContainerClient, Azure.Storage.DataMovement.Blobs.BlobStorageResourceContainerOptions options = null) { }
-        protected override bool CanProduceUri { get { throw null; } }
-        public override string Path { get { throw null; } }
         public override System.Uri Uri { get { throw null; } }
         protected override Azure.Storage.DataMovement.StorageResourceItem GetStorageResourceReference(string path) { throw null; }
         protected override System.Collections.Generic.IAsyncEnumerable<Azure.Storage.DataMovement.StorageResource> GetStorageResourcesAsync([System.Runtime.CompilerServices.EnumeratorCancellationAttribute] System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) { throw null; }
@@ -91,25 +87,11 @@ namespace Azure.Storage.DataMovement.Blobs
         public System.Collections.Generic.IDictionary<string, string> Tags { get { throw null; } set { } }
         public Azure.Storage.UploadTransferValidationOptions UploadTransferValidationOptions { get { throw null; } set { } }
     }
-    public partial class BlobStorageResourceProvider
-    {
-        internal BlobStorageResourceProvider() { }
-        public System.Threading.Tasks.Task<Azure.Storage.DataMovement.StorageResource> CreateResourceAsync(Azure.AzureSasCredential credential, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) { throw null; }
-        public System.Threading.Tasks.Task<Azure.Storage.DataMovement.StorageResource> CreateResourceAsync(Azure.Core.TokenCredential credential, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) { throw null; }
-        public System.Threading.Tasks.Task<Azure.Storage.DataMovement.StorageResource> CreateResourceAsync(System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) { throw null; }
-        public System.Threading.Tasks.Task<Azure.Storage.DataMovement.StorageResource> MakeResourceAsync(Azure.Storage.StorageSharedKeyCredential credential, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) { throw null; }
-    }
-    public static partial class BlobStorageResources
-    {
-        public static bool TryGetResourceProviders(Azure.Storage.DataMovement.DataTransferProperties info, out Azure.Storage.DataMovement.Blobs.BlobStorageResourceProvider sourceProvider, out Azure.Storage.DataMovement.Blobs.BlobStorageResourceProvider destinationProvider) { throw null; }
-    }
     public partial class BlockBlobStorageResource : Azure.Storage.DataMovement.StorageResourceItem
     {
         public BlockBlobStorageResource(Azure.Storage.Blobs.Specialized.BlockBlobClient blobClient, Azure.Storage.DataMovement.Blobs.BlockBlobStorageResourceOptions options = null) { }
-        protected override bool CanProduceUri { get { throw null; } }
         protected override long? Length { get { throw null; } }
         protected override long MaxChunkSize { get { throw null; } }
-        public override string Path { get { throw null; } }
         protected override string ResourceId { get { throw null; } }
         protected override Azure.Storage.DataMovement.DataTransferOrder TransferType { get { throw null; } }
         public override System.Uri Uri { get { throw null; } }
@@ -131,10 +113,8 @@ namespace Azure.Storage.DataMovement.Blobs
     public partial class PageBlobStorageResource : Azure.Storage.DataMovement.StorageResourceItem
     {
         public PageBlobStorageResource(Azure.Storage.Blobs.Specialized.PageBlobClient blobClient, Azure.Storage.DataMovement.Blobs.PageBlobStorageResourceOptions options = null) { }
-        protected override bool CanProduceUri { get { throw null; } }
         protected override long? Length { get { throw null; } }
         protected override long MaxChunkSize { get { throw null; } }
-        public override string Path { get { throw null; } }
         protected override string ResourceId { get { throw null; } }
         protected override Azure.Storage.DataMovement.DataTransferOrder TransferType { get { throw null; } }
         public override System.Uri Uri { get { throw null; } }